Add blockable metadata to pages and scaffold generator
Blockable metadata is a way to cache and store data on the blockable
record itself to avoid expensive database lookups when querying
information about a record that may come from the reord's blocks.

This commit also updates the page featured image metadata attribute to
look for an image url in the blockable metadata.