Fix bug in page callbacks hierarchy_changed? condition
hierarchy_changed? was always returning true because I had the if
syntax wrong. Previously these hierarchy and path methods were always
running when a page was saved, but now they will only run if
hierarchy_changed? is true. This is intended, but might mean there are
edge case bugs that I hadn't noticed previously that will need
to be dealt with.