Add even_of_kind? and odd_of_kind? methods to base block
This checks to see if a particular block is even or odd within instances
of that particular kind. A use case for this is if you have a layout
where a figure/text block change layouts during each instance of that
block, e.g. image on left when even, image on right when odd.