Don't initialize I18n settings when migrations need to run
This solves a problem that can occur when code in the settings model
relies on new changes to the database. Previously, the new setting code
would be  initialized and run before the migration, which caused the
migration to fail.

Now, the migration will run, and on the next app initialization the
settings will be initialized from your I18n files.